aedseo: Automated and Early Detection of Seasonal Epidemic Onset and Burden Levels

A powerful tool for automating the early detection of seasonal epidemic onsets in time series data. It offers the ability to estimate growth rates across consecutive time intervals, calculate the sum of cases (SoC) within those intervals, and estimate seasonal onsets within user defined seasons. With use of a disease-specific threshold it also offers the possibility to estimate seasonal onset of epidemics. Additionally it offers the ability to estimate burden levels for seasons based on historical data. It is aimed towards epidemiologists, public health professionals, and researchers seeking to identify and respond to seasonal epidemics in a timely fashion.

See also

Author

Maintainer: Lasse Engbo Christiansen lsec@ssi.dk (ORCID)

Authors:

Sofia Myrup Otero smot@ssi.dk (ORCID)

Kasper Schou Telkamp kasper.telkamp@gmail.com (ORCID)

Other contributors:

Rasmus Skytte Randløv rske@ssi.dk (ORCID) [reviewer]

Statens Serum Institut, SSI [copyright holder, funder]